requirements.txt
streamlit
python-dotenv
langchain-core
langchain-openai
langchain-community
langchain-chroma
chromadb